refactor: menus 패키지 수정 - menu 와 menugroup 패키지만 남김 - menu 에 'menu' 와 'menuProduct' 도메인 같이 위치
