+++ src/main/java/kitchenpos/menus/tobe/application/TobeMenuService.kt
@@ -4,7 +4,7 @@ import kitchenpos.menus.tobe.application.dto.CreateMenuRequest
 import kitchenpos.menus.tobe.domain.TobeMenuGroupRepository
 import kitchenpos.menus.tobe.domain.TobeMenuRepository
 import kitchenpos.menus.tobe.domain.menu.TobeMenu
-import kitchenpos.menus.tobe.domain.menuproduct.TobeMenuProduct
+import kitchenpos.menus.tobe.domain.menu.TobeMenuProduct
 import kitchenpos.shared.domain.DefaultProfanities
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
@@ -27,7 +27,14 @@ class TobeMenuService(
         }
         val name = request.name
         val price = request.price
-        val menuProducts = request.menuProducts.map { TobeMenuProduct(it.quantity, it.price, it.productId) }
+        val menuProducts =
+            request.menuProducts.map {
+                TobeMenuProduct(
+                    it.quantity,
+                    it.price,
+                    it.productId,
+                )
+            }
         val displayed = request.displayed
         val profanities = DefaultProfanities()
         val menu = TobeMenu(name, price, profanities, groupId, menuProducts)
