+++ src/main/java/kitchenpos/menus/tobe/domain/menu/TobeMenuProducts.java
@@ -1,7 +1,6 @@
 package kitchenpos.menus.tobe.domain.menu;
 
 import jakarta.persistence.*;
-import kitchenpos.menus.tobe.domain.menuproduct.TobeMenuProduct;
 
 import java.util.List;
