+++ src/test/java/kitchenpos/menus/tobe/domain/menu/TobeMenuProductsTest.java
@@ -1,6 +1,5 @@
 package kitchenpos.menus.tobe.domain.menu;
 
-import kitchenpos.menus.tobe.domain.menuproduct.TobeMenuProduct;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
